+++ src/v5/dispatcher.rs
@@ -4,19 +4,18 @@ use ntex_bytes::ByteString;
 use ntex_io::DispatchItem;
 use ntex_service::cfg::{Cfg, SharedCfg};
 use ntex_service::{self as service, Pipeline, Service, ServiceCtx, ServiceFactory};
+use ntex_util::services::buffer::{BufferService, BufferServiceError};
 use ntex_util::services::inflight::InFlightService;
-use ntex_util::services::{buffer::BufferService, buffer::BufferServiceError};
 use ntex_util::{HashMap, HashSet, future::join};
 
 use crate::error::{DecodeError, HandshakeError, MqttError, PayloadError, ProtocolError};
 use crate::payload::{Payload, PayloadStatus, PlSender};
 use crate::{MqttServiceConfig, types::QoS};
 
-use super::Session;
 use super::codec::{self, Decoded, DisconnectReasonCode, Encoded, Packet};
 use super::control::{Control, ControlAck};
 use super::publish::{Publish, PublishAck};
-use super::shared::{Ack, MqttShared};
+use super::{Session, shared::Ack, shared::MqttShared};
 
 /// MQTT 5 protocol dispatcher
 pub(super) fn factory<St, T, C, E>(
@@ -47,21 +46,27 @@ where
             join(factories.0.create(ses.clone()), factories.1.create(ses)).await;
 
         let publish = publish.map_err(|e| MqttError::Service(e.into()))?;
-        let control = control.map_err(|e| MqttError::Service(e.into()))?;
+        let control_unbuf = Pipeline::new(
+            control
+                .map_err(|e| MqttError::Service(e.into()))?
+                .map_err(|e| MqttError::Service(E::from(e))),
+        );
 
-        let control = BufferService::new(
-            16,
-            // limit number of in-flight messages
-            InFlightService::new(1, control),
-        )
-        .map_err(|err| match err {
-            BufferServiceError::Service(e) => MqttError::Service(E::from(e)),
-            BufferServiceError::RequestCanceled => {
-                MqttError::Handshake(HandshakeError::Disconnected(None))
-            }
-        });
+        let control = Pipeline::new(
+            BufferService::new(
+                16,
+                // limit number of in-flight messages
+                InFlightService::new(1, service::PipelineSvc::new(control_unbuf.clone())),
+            )
+            .map_err(|err| match err {
+                BufferServiceError::Service(e) => e,
+                BufferServiceError::RequestCanceled => {
+                    MqttError::Handshake(HandshakeError::Disconnected(None))
+                }
+            }),
+        );
 
-        Ok(Dispatcher::<_, _, E>::new(sink, publish, control, cfg))
+        Ok(Dispatcher::<_, _, _, E>::new(sink, publish, control, control_unbuf, cfg))
     })
 }
 
@@ -84,18 +89,20 @@ impl crate::inflight::SizedRequest for DispatchItem<Rc<MqttShared>> {
 }
 
 /// Mqtt protocol dispatcher
-pub(crate) struct Dispatcher<T, C: Service<Control<E>>, E> {
+pub(crate) struct Dispatcher<T, C: Service<Control<E>>, C2: Service<Control<E>>, E> {
     publish: T,
-    inner: Rc<Inner<C>>,
+    inner: Rc<Inner<C, C2>>,
     cfg: Cfg<MqttServiceConfig>,
     _t: marker::PhantomData<E>,
 }
 
-struct Inner<C> {
+struct Inner<C, C2> {
     control: Pipeline<C>,
+    control_unbuf: Pipeline<C2>,
     sink: Rc<MqttShared>,
     info: RefCell<PublishInfo>,
     payload: Cell<Option<PlSender>>,
+    stopped: Cell<bool>,
 }
 
 struct PublishInfo {
@@ -103,25 +110,34 @@ struct PublishInfo {
     aliases: HashMap<num::NonZeroU16, ByteString>,
 }
 
-impl<T, C, E> Dispatcher<T, C, E>
+impl<T, C, C2, E> Dispatcher<T, C, C2, E>
 where
     E: From<T::Error>,
     T: Service<Publish, Response = PublishAck>,
     PublishAck: TryFrom<T::Error, Error = E>,
     C: Service<Control<E>, Response = ControlAck, Error = MqttError<E>>,
+    C2: Service<Control<E>, Response = ControlAck, Error = MqttError<E>>,
 {
-    fn new(sink: Rc<MqttShared>, publish: T, control: C, cfg: Cfg<MqttServiceConfig>) -> Self {
+    fn new(
+        sink: Rc<MqttShared>,
+        publish: T,
+        control: Pipeline<C>,
+        control_unbuf: Pipeline<C2>,
+        cfg: Cfg<MqttServiceConfig>,
+    ) -> Self {
         Self {
             cfg,
             publish,
             inner: Rc::new(Inner {
                 sink,
+                control,
+                control_unbuf,
                 payload: Cell::new(None),
-                control: Pipeline::new(control),
                 info: RefCell::new(PublishInfo {
                     aliases: HashMap::default(),
                     inflight: HashSet::default(),
                 }),
+                stopped: Cell::new(false),
             }),
             _t: marker::PhantomData,
         }
@@ -132,7 +148,7 @@ where
     }
 }
 
-impl<C> Inner<C> {
+impl<C, C2> Inner<C, C2> {
     fn drop_payload<PErr>(&self, err: &PErr)
     where
         PErr: Clone,
@@ -144,12 +160,13 @@ impl<C> Inner<C> {
     }
 }
 
-impl<T, C, E> Service<DispatchItem<Rc<MqttShared>>> for Dispatcher<T, C, E>
+impl<T, C, C2, E> Service<DispatchItem<Rc<MqttShared>>> for Dispatcher<T, C, C2, E>
 where
     E: From<T::Error> + 'static,
     T: Service<Publish, Response = PublishAck> + 'static,
     PublishAck: TryFrom<T::Error, Error = E>,
     C: Service<Control<E>, Response = ControlAck, Error = MqttError<E>> + 'static,
+    C2: Service<Control<E>, Response = ControlAck, Error = MqttError<E>> + 'static,
 {
     type Response = Option<Encoded>;
     type Error = MqttError<E>;
@@ -159,8 +176,9 @@ where
         let result = if let Err(e) = res1 {
             if res2.is_err() {
                 Err(MqttError::Service(e.into()))
-            } else {
-                match self.inner.control.call(Control::error(e.into())).await {
+            } else if !self.inner.stopped.get() {
+                self.inner.stopped.set(true);
+                match self.inner.control_unbuf.call(Control::error(e.into())).await {
                     Ok(res) => {
                         if res.disconnect {
                             self.inner.sink.drop_sink();
@@ -169,30 +187,35 @@ where
                     }
                     Err(err) => Err(err),
                 }
+            } else {
+                res2
             }
         } else {
             res2
         };
 
-        if result.is_ok() {
-            if let Some(pl) = self.inner.payload.take() {
-                self.inner.payload.set(Some(pl.clone()));
-                if pl.ready().await != PayloadStatus::Ready {
-                    self.inner.sink.force_close();
-                }
+        if result.is_ok()
+            && let Some(pl) = self.inner.payload.take()
+        {
+            self.inner.payload.set(Some(pl.clone()));
+            if pl.ready().await != PayloadStatus::Ready {
+                self.inner.sink.force_close();
             }
         }
         result
     }
 
     fn poll(&self, cx: &mut Context<'_>) -> Result<(), Self::Error> {
-        if let Err(e) = self.publish.poll(cx) {
+        if let Err(e) = self.publish.poll(cx)
+            && !self.inner.stopped.get()
+        {
             let inner = self.inner.clone();
+            inner.stopped.set(true);
             ntex_rt::spawn(async move {
-                if let Ok(res) = inner.control.call(Control::error(e.into())).await {
-                    if res.disconnect {
-                        inner.sink.drop_sink();
-                    }
+                if let Ok(res) = inner.control_unbuf.call(Control::error(e.into())).await
+                    && res.disconnect
+                {
+                    inner.sink.drop_sink();
                 }
             });
         }
@@ -203,7 +226,7 @@ where
         log::trace!("{}: Shutdown v5 dispatcher", self.tag());
         self.inner.drop_payload(&PayloadError::Disconnected);
         self.inner.sink.drop_sink();
-        let _ = self.inner.control.call(Control::closed()).await;
+        let _ = self.inner.control.call(Control::shutdown()).await;
 
         self.publish.shutdown().await;
         self.inner.control.shutdown().await;
@@ -579,18 +602,19 @@ where
 }
 
 /// Publish service response future
-async fn publish_fn<'f, T, C, E>(
+async fn publish_fn<'f, T, C, C2, E>(
     publish: &T,
     pkt: Publish,
     packet_id: u16,
-    inner: &'f Inner<C>,
-    ctx: ServiceCtx<'f, Dispatcher<T, C, E>>,
+    inner: &'f Inner<C, C2>,
+    ctx: ServiceCtx<'f, Dispatcher<T, C, C2, E>>,
 ) -> Result<Option<Encoded>, MqttError<E>>
 where
     E: From<T::Error>,
     T: Service<Publish, Response = PublishAck>,
     PublishAck: TryFrom<T::Error, Error = E>,
     C: Service<Control<E>, Response = ControlAck, Error = MqttError<E>>,
+    C2: Service<Control<E>, Response = ControlAck, Error = MqttError<E>>,
 {
     let qos2 = pkt.qos() == QoS::ExactlyOnce;
     let ack = match ctx.call(publish, pkt).await {
@@ -629,17 +653,31 @@ where
     }
 }
 
-async fn control<C, E>(
+async fn control<C, C2, E>(
     pkt: Control<E>,
-    inner: &Inner<C>,
+    inner: &Inner<C, C2>,
     packet_id: u16,
 ) -> Result<Option<Encoded>, MqttError<E>>
 where
     C: Service<Control<E>, Response = ControlAck, Error = MqttError<E>>,
+    C2: Service<Control<E>, Response = ControlAck, Error = MqttError<E>>,
 {
-    let mut error = matches!(pkt, Control::Error(_) | Control::ProtocolError(_));
+    let mut stop = matches!(pkt, Control::Stop(_));
 
-    let result = match inner.control.call(pkt).await {
+    let result = if pkt.is_protocol() {
+        inner.control.call(pkt).await
+    } else {
+        if stop {
+            if !inner.stopped.get() {
+                inner.stopped.set(true)
+            } else {
+                return Ok(None);
+            }
+        }
+        inner.control_unbuf.call(pkt).await
+    };
+
+    let result = match result {
         Ok(result) => {
             if let Some(id) = num::NonZeroU16::new(packet_id) {
                 inner.info.borrow_mut().inflight.remove(&id);
@@ -649,24 +687,29 @@ where
         Err(err) => {
             inner.drop_payload(&PayloadError::Service);
 
-            // do not handle nested error
-            if error {
+            // do not handle nested stop
+            if stop {
                 inner.sink.drop_sink();
                 return Err(err);
             } else {
                 // handle error from control service
-                match err {
-                    MqttError::Service(err) => {
-                        error = true;
-                        inner.control.call(Control::error(err)).await?
+                if !inner.stopped.get() {
+                    inner.stopped.set(true);
+
+                    if let MqttError::Service(err) = err {
+                        stop = true;
+                        inner.control_unbuf.call(Control::error(err)).await?
+                    } else {
+                        return Err(err);
                     }
-                    _ => return Err(err),
+                } else {
+                    return Err(err);
                 }
             }
         }
     };
 
-    let response = if error {
+    let response = if stop {
         if let Some(pkt) = result.packet {
             let _ = inner.sink.encode_packet(pkt);
         }
@@ -684,8 +727,11 @@ where
 
 #[cfg(test)]
 mod tests {
+    use std::num::NonZeroU16;
+
+    use ntex_bytes::Bytes;
     use ntex_io::{Io, testing::IoTest};
-    use ntex_service::{cfg::SharedCfg, fn_service};
+    use ntex_service::fn_service;
     use ntex_util::future::{Ready, lazy};
 
     use super::*;
@@ -707,16 +753,15 @@ mod tests {
         let io = Io::new(IoTest::create().0, SharedCfg::new("DBG"));
         let codec = codec::Codec::default();
         let shared = Rc::new(MqttShared::new(io.get_ref(), codec, Default::default()));
+        let control = Pipeline::new(fn_service(|_| {
+            Ready::Ok::<_, MqttError<TestError>>(ControlAck { packet: None, disconnect: false })
+        }));
 
-        let disp = Pipeline::new(Dispatcher::<_, _, _>::new(
+        let disp = Pipeline::new(Dispatcher::<_, _, _, _>::new(
             shared.clone(),
             fn_service(|p: Publish| Ready::Ok::<_, TestError>(p.ack())),
-            fn_service(|_| {
-                Ready::Ok::<_, MqttError<TestError>>(ControlAck {
-                    packet: None,
-                    disconnect: false,
-                })
-            }),
+            control.clone(),
+            control,
             Default::default(),
         ));
 
@@ -736,5 +781,48 @@ mod tests {
         disp.call(DispatchItem::WBackPressureDisabled).await.unwrap();
         assert!(lazy(|cx| rx.poll_recv(cx).is_ready()).await);
         assert!(!lazy(|cx| rx2.poll_recv(cx).is_ready()).await);
+    }
+
+    #[ntex::test]
+    async fn control_stop_once_on_service_error() {
+        let io = Io::new(IoTest::create().0, SharedCfg::new("DBG"));
+        let codec = codec::Codec::default();
+        let shared = Rc::new(MqttShared::new(io.get_ref(), codec, Default::default()));
+        let counter = Rc::new(Cell::new(0));
+        let counter2 = counter.clone();
+        let control = Pipeline::new(fn_service(async move |msg| {
+            if matches!(msg, Control::Stop(_)) {
+                counter2.set(counter2.get() + 1);
+            }
+            Ok::<_, MqttError<TestError>>(ControlAck { packet: None, disconnect: false })
+        }));
+
+        let disp = Pipeline::new(Dispatcher::<_, _, _, _>::new(
+            shared.clone(),
+            fn_service(async |_: Publish| Err(TestError)),
+            control.clone(),
+            control,
+            Default::default(),
+        ));
+
+        disp.call(DispatchItem::Item(Decoded::Publish(
+            codec::Publish {
+                dup: false,
+                retain: false,
+                qos: codec::QoS::AtLeastOnce,
+                topic: ByteString::from("test"),
+                packet_id: Some(NonZeroU16::new(1).unwrap()),
+                payload_size: 0,
+                properties: Default::default(),
+            },
+            Bytes::new(),
+            0,
+        )))
+        .await
+        .unwrap();
+
+        disp.call(DispatchItem::Disconnect(None)).await.unwrap();
+
+        assert_eq!(counter.get(), 1);
     }
 }